Prediction/requirements.txt
biopython==1.81
torch==2.0.1
numpy
pandas